refactor ocds.py to improve codeclimate
