+++ src/openprocurement/api/models/registry_models/ocds.py
@@ -4,10 +4,8 @@ from uuid import uuid4
 from schematics.types import (
     StringType,
     FloatType,
-    URLType,
     BooleanType,
     BaseType,
-    EmailType,
     MD5Type,
 )
 from schematics.exceptions import ValidationError
@@ -23,11 +21,16 @@ from openprocurement.api.constants import (
     DEFAULT_ITEM_CLASSIFICATION,
     ITEM_CLASSIFICATIONS,
     DOCUMENT_TYPES,
-    IDENTIFIER_CODES,
     DEBTOR_TYPES
 )
 from openprocurement.api.utils import get_now, serialize_document_url
-
+from openprocurement.api.models.models import (
+    ContactPoint,
+    Classification,
+    Identifier,
+    Address,
+    Location
+)
 from openprocurement.api.models.schematics_extender import (
     Model,
     IsoDateTimeType,
@@ -37,41 +40,38 @@ from .roles import document_roles, organization_roles
 
 
 # OCDS Building Blocks.
-# More info: http://standard.open-contracting.org/latest/en/getting_started/building_blocks
+# More info:
+# http://standard.open-contracting.org/latest/en/getting_started/building_blocks
 
 class BasicValue(Model):
     amount = FloatType(required=True, min_value=0)  # Amount as a number.
-    currency = StringType(required=True, max_length=3, min_length=3)  # The currency in 3-letter ISO 4217 format.
+    # The currency in 3-letter ISO 4217 format.
+    currency = StringType(required=True, max_length=3, min_length=3)
 
 
 class Value(BasicValue):
     valueAddedTaxIncluded = BooleanType(required=True, default=True)
     # The currency in 3-letter ISO 4217 format.
-    currency = StringType(required=True, default=DEFAULT_CURRENCY, max_length=3, min_length=3)
+    currency = StringType(required=True, default=DEFAULT_CURRENCY,
+                          max_length=3, min_length=3)
 
 
 class ValueUAH(BasicValue):
     # The currency in 3-letter ISO 4217 format.
-    currency = StringType(required=True, choices=[u'UAH'], max_length=3, min_length=3)
-
-
-class Classification(Model):
-    scheme = StringType(required=True)  # The classification scheme for the goods
-    id = StringType(required=True)  # The classification ID from the Scheme used
-    description = StringType(required=True)  # A description of the goods, services to be provided.
-    description_en = StringType()
-    description_ru = StringType()
-    uri = URLType()
+    currency = StringType(required=True, choices=[u'UAH'],
+                          max_length=3, min_length=3)
 
 
 class ItemClassification(Classification):
-    scheme = StringType(required=True, default=DEFAULT_ITEM_CLASSIFICATION, choices=ITEM_CLASSIFICATIONS.keys())
+    scheme = StringType(required=True, default=DEFAULT_ITEM_CLASSIFICATION,
+                        choices=ITEM_CLASSIFICATIONS.keys())
     id = StringType(required=True)
 
     def validate_id(self, data, code):
         available_codes = ITEM_CLASSIFICATIONS.get(data.get('scheme'), [])
         if code not in available_codes:
-            raise ValidationError(BaseType.MESSAGES['choices'].format(unicode(available_codes)))
+            raise ValidationError(
+                BaseType.MESSAGES['choices'].format(unicode(available_codes)))
 
 
 class BaseUnit(Model):
@@ -94,23 +94,6 @@ class Unit(BaseUnit):
     value = ModelType(Value)
 
 
-class Address(Model):
-
-    streetAddress = StringType()
-    locality = StringType()
-    region = StringType()
-    postalCode = StringType()
-    countryName = StringType(required=True)
-    countryName_en = StringType()
-    countryName_ru = StringType()
-
-
-class Location(Model):
-    latitude = BaseType(required=True)
-    longitude = BaseType(required=True)
-    elevation = BaseType()
-
-
 class Document(Model):
     class Options:
         roles = document_roles
@@ -128,7 +111,8 @@ class Document(Model):
     format = StringType(required=True, regex='^[-\w]+/[-\.\w\+]+$')
     url = StringType(required=True)  # Link to the document or attachment.
     datePublished = IsoDateTimeType(default=get_now)
-    dateModified = IsoDateTimeType(default=get_now)  # Date that the document was last dateModified
+    # Date that the document was last dateModified
+    dateModified = IsoDateTimeType(default=get_now)
     language = StringType()
     relatedItem = MD5Type()
     author = StringType()
@@ -153,25 +137,18 @@ class Document(Model):
         return self
 
 
-class Identifier(Model):
-    # The scheme that holds the unique identifiers used to identify the item being identified.
-    scheme = StringType(required=True, choices=IDENTIFIER_CODES)
-    id = BaseType(required=True)  # The identifier of the organization in the selected scheme.
-    legalName = StringType()  # The legally registered name of the organization.
-    legalName_en = StringType()
-    legalName_ru = StringType()
-    uri = URLType()  # A URI to identify the organization.
-
-
 class Item(Model):
     """A good, service, or work to be contracted."""
     id = StringType(required=True, min_length=1, default=lambda: uuid4().hex)
-    description = StringType(required=True)  # A description of the goods, services to be provided.
+    # A description of the goods, services to be provided.
+    description = StringType(required=True)
     description_en = StringType()
     description_ru = StringType()
     classification = ModelType(ItemClassification)
-    additionalClassifications = ListType(ModelType(Classification), default=list())
-    unit = ModelType(Unit)  # Description of the unit which the good comes in e.g. hours, kilograms
+    additionalClassifications = ListType(
+        ModelType(Classification), default=list())
+    # Description of the unit which the good comes in e.g. hours, kilograms
+    unit = ModelType(Unit)
     quantity = DecimalType()  # The number of units required
     address = ModelType(Address)
     location = ModelType(Location)
@@ -180,22 +157,11 @@ class Item(Model):
     def validate_schema_properties(self, data, new_schema_properties):
         """ Raise validation error if code in schema_properties mismatch
             with classification id """
-        if new_schema_properties and not data['classification']['id'].startswith(new_schema_properties['code']):
-            raise ValidationError("classification id mismatch with schema_properties code")
-
-
-class ContactPoint(Model):
-    name = StringType(required=True)
-    name_en = StringType()
-    name_ru = StringType()
-    email = EmailType()
-    telephone = StringType()
-    faxNumber = StringType()
-    url = URLType()
-
-    def validate_email(self, data, value):
-        if not value and not data.get('telephone'):
-            raise ValidationError(u"telephone or email should be present")
+        mismatch = not data['classification']['id']\
+            .startswith(new_schema_properties['code'])
+        if new_schema_properties and mismatch:
+            raise ValidationError(
+                "classification id mismatch with schema_properties code")
 
 
 class Organization(Model):
